pyproject.toml
[tool.poetry]
name = "beatcovid-geoprocess"
version = "1.0.0"
description = ""
authors = ["beatcovi19now <git@beatcovid19now.org>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
pandas = "^1.0.3"
python-dotenv = "^0.13.0"
requests-cache = "^0.5.2"
pymongo = "^3.10.1"
geopandas = "^0.7.0"
shapely = "^1.7.0"
huey = "^2.2.0"
redis = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "19.10b0"
flake8 = "^3.7.8"
isort = "^4.3.21"
pylint = "^2.2"
rope = "^0.16.0"
pylint-django = "^2.0.14"
ipython = "^7.13.0"
jupyter = "^1.0.0"

[tool.black]
line-length = 90
target_version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
